Move mapStateToProps up to routing component. Make sure invalid urls get redirected.
